switching to trying 8259 pic instead of ioapic
